requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
streamlit==1.35.0
numpy
